backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
PyPDF2>=3.0.0
numpy>=1.20.0
requests>=2.28.0
sentence-transformers>=2.2.2
rank-bm25>=0.2.2
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
openai>=1.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=9.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
tiktoken>=0.5.1 